Refactor collection links to utilities
